# Add a mapping of Zlib to Libdeflate compression levels for BGZF.