@jkbonfield 2022-08-01
Also curious is how BCF is a much poorer format to compress for single sample data than VCF (but better on multi-sample).

Compression of BCF is 15% and 12% larger than VCF at bgzf level 7 for the single-sample data, but VCF.gz is 19% larger than BCF for the 4000-sample data.  It's not clear why this is so.   (Speed wise BCF was 30-90% faster to encode, but likely better on decode.)